Move errors from internal to pack package
