+++ internal/commands/build.go
@@ -41,11 +41,11 @@ type BuildFlags struct {
 func validateBuildFlags(flags BuildFlags, logger logging.Logger, cfg config.Config, packClient PackClient) error {
 	if flags.Builder == "" {
 		suggestSettingBuilder(logger, packClient)
-		return NewSoftError()
+		return pack.NewSoftError()
 	}
 
 	if flags.Registry != "" && !cfg.Experimental {
-		return NewExperimentError("Support for buildpack registries is currently experimental.")
+		return pack.NewExperimentError("Support for buildpack registries is currently experimental.")
 	}
 
 	return nil
