--- internal/commands/soft_error.go
@@ -1,12 +1,0 @@
-package commands
-
-type SoftError struct {
-}
-
-func NewSoftError() SoftError {
-	return SoftError{}
-}
-
-func (se SoftError) Error() string {
-	return ""
-}
